requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Werkzeug==2.3.6
bleach==6.0.0
Flask-WTF==1.1.1
tailwindcss==0.1.0
pytz==2024.1
daisyui==4.6.0
engineio==4.5.1
gevent==21.12.0
gevent-websocket==0.10.1